+++ src/constants/ProjectFileContent.ts
@@ -1,5 +1,5 @@
-
-
+/* eslint-disable no-unused-vars */
+// eslint-disable-next-line no-shadow
 export enum ProjectFileContent {
   init = 'init',
   initScenario = 'init-scenario',