+++ src/services/CliActions.ts
@@ -1,14 +1,13 @@
 import { ServeConfig } from './../models/ServeConfig';
-import { readFileSync } from 'fs';
+import { readFileSync, writeFileSync } from 'fs';
 import { Folders } from './Folders';
-import { commands, Progress, ProgressLocation, Range, Uri, window, workspace } from "vscode";
-import { Commands } from "../constants";
-import { SolutionAddResult, Subscription } from "../models";
-import { Extension } from "./Extension";
-import { CliExecuter } from "./CliCommandExecuter";
-import { Notifications } from "./Notifications";
+import { commands, Progress, ProgressLocation, Uri, window, workspace } from 'vscode';
+import { Commands } from '../constants';
+import { SolutionAddResult, Subscription } from '../models';
+import { Extension } from './Extension';
+import { CliExecuter } from './CliCommandExecuter';
+import { Notifications } from './Notifications';
 import { basename, join } from 'path';
-import { writeFileSync } from 'fs';
 import { EnvironmentInformation } from './EnvironmentInformation';
 import { AuthProvider } from '../providers/AuthProvider';
 
@@ -17,7 +16,7 @@ export class CliActions {
 
   public static registerCommands() {
     const subscriptions: Subscription[] = Extension.getInstance().subscriptions;
-    
+
     subscriptions.push(
       commands.registerCommand(Commands.upgradeSolution, CliActions.upgrade)
     );
@@ -36,7 +35,7 @@ export class CliActions {
    * Get the root SPO URL
    */
   public static async appCatalogUrlGet() {
-    const data = (await CliExecuter.execute("spo tenant appcatalogurl get", "json")).stdout || undefined;
+    const data = (await CliExecuter.execute('spo tenant appcatalogurl get', 'json')).stdout || undefined;
     EnvironmentInformation.appCatalogUrl = data ? JSON.parse(data) : undefined;
     return EnvironmentInformation.appCatalogUrl;
   }
@@ -53,21 +52,22 @@ export class CliActions {
 
     await window.withProgress({
       location: ProgressLocation.Notification,
-      title: `Generating the upgrade steps...`,
+      title: 'Generating the upgrade steps...',
       cancellable: true
+      // eslint-disable-next-line no-unused-vars
     }, async (progress: Progress<{ message?: string; increment?: number }>) => {
       try {
-        const result = await CliExecuter.execute("spfx project upgrade", "md");
-        
+        const result = await CliExecuter.execute('spfx project upgrade', 'md');
+
         if (result.stdout) {
           // Create a file to allow the Markdown preview to correctly open the linked/referenced files
-          const filePath = join(wsFolder?.uri.fsPath || "", "spfx.upgrade.md");
+          const filePath = join(wsFolder?.uri.fsPath || '', 'spfx.upgrade.md');
           writeFileSync(filePath, result.stdout);
-          await commands.executeCommand("markdown.showPreview", Uri.file(filePath));
+          await commands.executeCommand('markdown.showPreview', Uri.file(filePath));
         } else if (result.stderr) {
           Notifications.error(result.stderr);
         }
-      } catch(e: any) {
+      } catch (e: any) {
         const message = e?.error?.message;
         Notifications.error(message);
       }
@@ -86,21 +86,22 @@ export class CliActions {
 
     await window.withProgress({
       location: ProgressLocation.Notification,
-      title: `Validating the current solution...`,
+      title: 'Validating the current solution...',
       cancellable: true
+      // eslint-disable-next-line no-unused-vars
     }, async (progress: Progress<{ message?: string; increment?: number }>) => {
       try {
-        const result = await CliExecuter.execute("spfx project doctor", "md");
-        
+        const result = await CliExecuter.execute('spfx project doctor', 'md');
+
         if (result.stdout) {
           // Create a file to allow the Markdown preview to correctly open the linked/referenced files
-          const filePath = join(wsFolder?.uri.fsPath || "", "spfx.validate.md");
+          const filePath = join(wsFolder?.uri.fsPath || '', 'spfx.validate.md');
           writeFileSync(filePath, result.stdout);
-          await commands.executeCommand("markdown.showPreview", Uri.file(filePath));
+          await commands.executeCommand('markdown.showPreview', Uri.file(filePath));
         } else if (result.stderr) {
           Notifications.error(result.stderr);
         }
-      } catch(e: any) {
+      } catch (e: any) {
         const message = e?.error?.message;
         Notifications.error(message);
       }
@@ -115,25 +116,25 @@ export class CliActions {
     const account = await authInstance.getAccount();
 
     if (!account) {
-      Notifications.error("You must be logged in to deploy a solution.");
+      Notifications.error('You must be logged in to deploy a solution.');
       return;
     }
 
     if (!file) {
-      const sppkgFiles = await workspace.findFiles(`**/*.sppkg`, `**/node_modules/**`);
+      const sppkgFiles = await workspace.findFiles('**/*.sppkg', '**/node_modules/**');
       if (sppkgFiles.length <= 0) {
-        Notifications.error("No sppkg files found in the workspace");
+        Notifications.error('No sppkg files found in the workspace');
         return;
       }
 
       if (sppkgFiles.length > 1) {
-        Notifications.error("Multiple sppkg files found in the workspace");
+        Notifications.error('Multiple sppkg files found in the workspace');
 
         const answer = await window.showQuickPick(sppkgFiles.map(f => basename(f.fsPath)), {
-          placeHolder: "Select the sppkg file to deploy",
+          placeHolder: 'Select the sppkg file to deploy',
           ignoreFocusOut: true,
           canPickMany: false,
-          title: "Select the sppkg file to deploy"
+          title: 'Select the sppkg file to deploy'
         });
 
         if (!answer) {
@@ -151,7 +152,7 @@ export class CliActions {
     }
 
     if (!EnvironmentInformation.appCatalogUrl) {
-      Notifications.error("We haven't been able to find an app catalog URL. Make sure your environment has an app catalog site.");
+      Notifications.error('We haven\'t been able to find an app catalog URL.Make sure your environment has an app catalog site.');
       return;
     }
 
@@ -159,10 +160,11 @@ export class CliActions {
       location: ProgressLocation.Notification,
       title: `Deploying the ${basename(file.fsPath)} solution. Check [output window](command:${Commands.showOutputChannel}) to follow the progress.`,
       cancellable: false
+      // eslint-disable-next-line no-unused-vars
     }, async (progress: Progress<{ message?: string; increment?: number }>) => {
       try {
-        const addResult = await CliExecuter.execute(`spo app add`, "json", { filePath: file?.fsPath, appCatalogUrl: EnvironmentInformation.appCatalogUrl, overwrite: true });
-        
+        const addResult = await CliExecuter.execute('spo app add', 'json', { filePath: file?.fsPath, appCatalogUrl: EnvironmentInformation.appCatalogUrl, overwrite: true });
+
         if (addResult.stderr) {
           Notifications.error(addResult.stderr);
           return;
@@ -170,28 +172,28 @@ export class CliActions {
 
         const data: SolutionAddResult = JSON.parse(addResult.stdout);
         if (!data.UniqueId) {
-          Notifications.error("We haven't been able to find the unique ID of the solution. Make sure the solution was added correctly.");
+          Notifications.error('We haven\'t been able to find the unique ID of the solution.Make sure the solution was added correctly.');
         }
 
         // Check if skip feature deployment
-        const packageSolution = await workspace.findFiles(`**/config/package-solution.json`, `**/node_modules/**`);
+        const packageSolution = await workspace.findFiles('**/config/package-solution.json', '**/node_modules/**');
         let shouldSkipFeatureDeployment = false;
         if (packageSolution.length > 0) {
           const packageSolutionFile = packageSolution[0];
-          const packageSolutionContents = readFileSync(packageSolutionFile.fsPath, "utf8");
+          const packageSolutionContents = readFileSync(packageSolutionFile.fsPath, 'utf8');
           const packageSolutionData = JSON.parse(packageSolutionContents);
           shouldSkipFeatureDeployment = packageSolutionData.solution.skipFeatureDeployment;
         }
 
-        const deployResult = await CliExecuter.execute(`spo app deploy`, "json", { id: data.UniqueId, appCatalogUrl: EnvironmentInformation.appCatalogUrl, skipFeatureDeployment: shouldSkipFeatureDeployment });
+        const deployResult = await CliExecuter.execute('spo app deploy', 'json', { id: data.UniqueId, appCatalogUrl: EnvironmentInformation.appCatalogUrl, skipFeatureDeployment: shouldSkipFeatureDeployment });
 
         if (deployResult.stderr) {
           Notifications.error(addResult.stderr);
           return;
         }
 
-        Notifications.info("The solution has been deployed successfully.");
-      } catch(e: any) {
+        Notifications.info('The solution has been deployed successfully.');
+      } catch (e: any) {
         const message = e?.error?.message;
         Notifications.error(message);
       }
@@ -207,19 +209,19 @@ export class CliActions {
       return;
     }
 
-    const serveFiles = await workspace.findFiles("config/serve.json", "**/node_modules/**");
-    let serveFile = serveFiles && serveFiles.length > 0 ? serveFiles[0] : null;
+    const serveFiles = await workspace.findFiles('config/serve.json', '**/node_modules/**');
+    const serveFile = serveFiles && serveFiles.length > 0 ? serveFiles[0] : null;
 
     if (!serveFile) {
       return;
     }
 
-    const serveFileContents = readFileSync(serveFile.fsPath, "utf8");
+    const serveFileContents = readFileSync(serveFile.fsPath, 'utf8');
     const serveFileData: ServeConfig = JSON.parse(serveFileContents);
     const configNames = Object.keys(serveFileData.serveConfigurations);
 
     const answer = await window.showQuickPick(configNames, {
-      title: "Select the configuration to serve",
+      title: 'Select the configuration to serve',
       ignoreFocusOut: true
     });