+++ src/webview/view/components/gallery/List.tsx
@@ -4,6 +4,7 @@ import { useEffect, useState } from 'react';
 import { Sample } from '../../../../models';
 import { Card } from './Card';
 
+
 export interface IListProps {
   items: Sample[];
 }
@@ -29,15 +30,15 @@ export const List: React.FunctionComponent<IListProps> = ({ items }: React.Props
     setPagedItems(getNewSet(page));
   }, [page, items]);
 
-  const getNewSet = (page: number): Sample[] => items.slice(page * NR_OF_ITEMS, page * NR_OF_ITEMS + NR_OF_ITEMS);
+  const getNewSet = (pageNumber: number): Sample[] => items.slice(pageNumber * NR_OF_ITEMS, page * NR_OF_ITEMS + NR_OF_ITEMS);
 
   if (!pagedItems || pagedItems.length === 0) {
     return null;
   }
 
   return (
-    <div className={`w-full flex justify-between flex-col flex-grow max-w-7xl mx-auto pt-6`}>
-      <ul role="list" className={`grid grid-cols-2 gap-x-4 gap-y-8 sm:grid-cols-3 sm:gap-x-6 lg:grid-cols-4 xl:gap-x-8`}>
+    <div className={'w-full flex justify-between flex-col flex-grow max-w-7xl mx-auto pt-6'}>
+      <ul role="list" className={'grid grid-cols-2 gap-x-4 gap-y-8 sm:grid-cols-3 sm:gap-x-6 lg:grid-cols-4 xl:gap-x-8'}>
         {
           pagedItems.map((item, index) => (
             <Card key={`${index}-${item.name}`} item={item} />
@@ -45,13 +46,13 @@ export const List: React.FunctionComponent<IListProps> = ({ items }: React.Props
         }
       </ul>
 
-      <div className={`flex justify-center items-center space-x-4 mt-8`}>
+      <div className={'flex justify-center items-center space-x-4 mt-8'}>
         <VSCodeButton
           onClick={() => setPage(page - 1)}
           disabled={page === 0}>
           Previous
         </VSCodeButton>
-        
+
         <VSCodeButton
           onClick={() => setPage(page + 1)}
           disabled={page * NR_OF_ITEMS + NR_OF_ITEMS >= items.length}>